# api/swagger/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/admin/creators-list": {
      "get": {
        "operationId": "CreatorListController_findAll",
        "parameters": [
          {
            "name": "page",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "default": 1,
              "example": 1,
              "type": "number"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "maximum": 100,
              "default": 20,
              "example": 20,
              "type": "number"
            }
          },
          {
            "name": "search",
            "required": false,
            "in": "query",
            "description": "Search by name or email, case-insensitive",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "status",
            "required": false,
            "in": "query",
            "schema": {
              "$ref": "#/components/schemas/EAccountStatus"
            }
          },
          {
            "name": "sortBy",
            "required": false,
            "in": "query",
            "schema": {
              "$ref": "#/components/schemas/EAccountSortField"
            }
          },
          {
            "name": "sortOrder",
            "required": false,
            "in": "query",
            "schema": {
              "$ref": "#/components/schemas/ESortOrder"
            }
          },
          {
            "name": "emailVerified",
            "required": false,
            "in": "query",
            "description": "true = email đã xác minh, false = chưa",
            "schema": {
              "type": "boolean"
            }
          },
          {
            "name": "createdFrom",
            "required": false,
            "in": "query",
            "schema": {
              "format": "date",
              "example": "2026-01-01",
              "type": "string"
            }
          },
          {
            "name": "createdTo",
            "required": false,
            "in": "query",
            "schema": {
              "format": "date",
              "example": "2026-12-31",
              "type": "string"
            }
          },
          {
            "name": "address",
            "required": false,
            "in": "query",
            "description": "Match against the address in the role profile table",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "gender",
            "required": false,
            "in": "query",
            "description": "1 = male, 2 = female, 3 = other",
            "schema": {
              "$ref": "#/components/schemas/GendersEnum"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AccountFilterResponseDto"
                }
              }
            }
          },
          "401": {
            "description": "Token is missing, invalid or expired"
          },
          "403": {
            "description": "Requires admin role"
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "List creator accounts, paginated",
        "tags": [
          "Admin"
        ]
      }
    },
    "/admin/creators-list/{id}": {
      "delete": {
        "operationId": "CreatorListController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "message": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Token is missing, invalid or expired"
          },
          "403": {
            "description": "Requires admin role"
          },
          "404": {
            "description": "Creator profile not found"
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "Delete a creator profile",
        "tags": [
          "Admin"
        ]
      }
    },
...
  },
  "info": {
    "title": "CRM KOC API",
    "description": "REST API for the CRM KOC system",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "access-token": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "ERole": {
        "type": "string",
        "enum": [
          "admin",
          "brand",
          "creator"
        ]
      },
      "EAccountStatus": {
        "type": "number",
        "enum": [
          1,
          2,
          3,
          4
        ]
      },
...
      "EAccountSortField": {
        "type": "string",
        "enum": [
          "createdAt",
          "updatedAt",
          "name",
          "email",
          "status"
        ]
      },
      "ESortOrder": {
        "type": "string",
        "enum": [
          "ASC",
          "DESC"
        ]
      },
      "AccountFilterItemDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "name": {
            "type": "string",
            "maxLength": 255
          },
          "email": {
            "type": "string",
            "format": "email"
          },
          "phone": {
            "type": "string",
            "nullable": true,
            "example": "+84901111111"
          },
          "accountRole": {
            "allOf": [
              {
                "$ref": "#/components/schemas/ERole"
              }
            ]
          },
          "status": {
            "allOf": [
              {
                "$ref": "#/components/schemas/EAccountStatus"
              }
            ]
          },
          "createdAt": {
            "format": "date-time",
            "type": "string"
          }
        },
        "required": [
          "id",
          "name",
          "email",
          "phone",
          "accountRole",
          "status",
          "createdAt"
        ]
      },
      "AccountFilterResponseDto": {
        "type": "object",
        "properties": {
          "data": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/AccountFilterItemDto"
            }
          },
          "total": {
            "type": "number",
            "example": 3,
            "description": "Total rows matching the filter"
          },
          "page": {
            "type": "number",
            "example": 1
          },
          "limit": {
            "type": "number",
            "example": 20
          },
          "totalPages": {
            "type": "number",
            "example": 1
          }
        },
        "required": [
          "data",
          "total",
          "page",
          "limit",
          "totalPages"
        ]
      },
      "GendersEnum": {
        "type": "string",
        "enum": [
          1,
          2,
          3
        ]
      },
...
    }
  }
}